requirements.txt
pytorch
multiprocessing
numpy
matplotlib
pandas
